delete works

+++ src/main/java/ui/controller/Delete.java
@@ -11,9 +11,18 @@ public class Delete extends RequestHandler {
 
     @Override
     public String handleRequest(HttpServletRequest request, HttpServletResponse response) {
-        int id = Integer.parseInt(request.getParameter("id"));
-        System.out.println("delete "+id);
-        service.delete(id);
-        return "OK";
+        String id_str = null;
+        try {
+            ObjectMapper objectMapper = new ObjectMapper();
+            id_str = objectMapper.readValue(request.getInputStream(), String.class);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        if (id_str != null) {
+            int id = Integer.parseInt(id_str);
+            service.delete(id);
+            return "OK";
+        }
+        return "NOK";
     }
 }
